Changed to PCG minimal random generator instead of linux rand

+++ worker/math.h
@@ -17,5 +17,6 @@ Unt round_up_to_power_of_2(Unt n);
 #define MIN(x, y) (((x) <= (y)) ? (x) : (y))
 #endif
 
-Double random_int(Int lower, Int upper);
+Int random_int(Int lower, Int upper);
+
 #endif
